Consolidate NVML tests
